refactoring/authProvider: provider moved into contexts, fetch separated, import paths updated
