+++ src/utils/routers/PrivateRoute.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Navigate } from "react-router-dom";
-import { useAuthContext } from "../AuthProvider";
+import { useAuthContext } from "../contexts/AuthProvider";
 
 const PrivateRoute = ({ children }) => {
   const { isLoggedIn } = useAuthContext();
